fix: resolve all 70+ test build errors to clean build
- Fix PaymentDocument.Amount -> PaymentDocument.Request.Amount
- Fix PaymentSagaState.PaymentPaymentRequest -> PaymentRequest
- Fix all PaymentCommand.Request initializers across 9 test files
- Add missing usings for Worker namespaces (Commands, Events, Sagas, Services, Dtos)
- Add PaymentEnriched.EnrichmentData property
- Fix double->long conversion in DockerIntegrationTests
- Add Xunit.SkippableFact package
- Fix MetricsAndConfigTests operator ambiguity with parentheses
- Source + tests: 0 errors, 0 warnings

+++ tests/PaymentService.WorkerService.IntegrationTests/DockerIntegrationTests.cs
@@ -222,7 +222,7 @@ public class DockerIntegrationTests
         sagas.Should().HaveCount(paymentCount);
         sagas.Should().AllSatisfy(s => s.State.Status.Should().Be("Settled"));
         dlq.PublishedEvents.Should().BeEmpty();
-        sw.ElapsedMilliseconds.Should().BeLessThan(TimeSpan.FromSeconds(10).TotalMilliseconds,
+        sw.ElapsedMilliseconds.Should().BeLessThan((long)TimeSpan.FromSeconds(10).TotalMilliseconds,
             "50 saga flows should complete within 10 seconds");
     }
